Refactor Retry-After parsing into rate limit error
